in progress

+++ Inc/stepperMotors.h
@@ -10,10 +10,10 @@
 void stepperABEnable(void);
 void stepperABDisable(void);
 void stepperABInit(void);
-void stepperASetSpeed(uint8_t setpoint);
-void stepperBSetSpeed(uint8_t setpoint);
+void stepperASetSpeed(uint16_t setpoint);
+void stepperBSetSpeed(uint16_t setpoint);
 void stepperASetDir(bool dir);
 void stepperBSetDir(bool dir);
-void stepperABMove(uint8_t setpoint, bool dir);
+void stepperABMove(uint16_t setpoint, bool dir);
 
 #endif /* STEPPERMOTORS_H_ */
